Add interactive list to the todo
