+++ src/modules/functionality.js
@@ -1,0 +1,10 @@
+const setState = (tasks, checkbox, index) => {
+    if (checkbox.checked === true) {
+      tasks[index].checked = true;
+    } else {
+      tasks[index].checked = false;
+    }
+  };
+  
+  export default setState;
+
