--- src/modules/getStates.js
@@ -1,9 +1,0 @@
-const setState = (tasks, checkbox, index) => {
-  if (checkbox.checked === true) {
-    tasks[index].checked = true;
-  } else {
-    tasks[index].checked = false;
-  }
-};
-
-export default setState;
